🌳 Windows' Temp Directory - 🐛 Hidden Files? More like, *shhh* hidden files! 😏
Hark! The tale of our intrepid temp directory, where files can be as elusive as a leprechaun's pot o' gold. Our hero tried to create some hidden treasures (`.hidden/subdir/deep.txt`), but ol' Windows said, "Not on my watch!" 💥🛑

But fear not, for our steadfast test has adapted like the mighty oak! It now checks if visible.txt is still there, even when hidden files play hide-and-seek. And if they're found (on non-Windows systems), well, that's just ✨ magic ✨, isn't it? 🎩🐇

In other news, we've added a handy debug printout for those pesky Windows temp directory issues. So, if you find yourself in the depths of debugging, remember to keep your eyes peeled for this helpful hint!

Signed with a smile,
Aye (aye@8b.is)